Organize subprocess types and runner files
