+++ Sources/ClawSubprocess/Lock/InstanceLock.swift
