+++ Sources/ClawSubprocess/Runner/SubprocessRunning.swift
@@ -1,0 +1,3 @@
+package protocol SubprocessRunning: Sendable {
+  func run(_ command: SubprocessCommand) async -> SubprocessResult
+}
