+++ Sources/ClawSubprocess/Runner/SwiftSubprocessRunner.swift
@@ -9,78 +9,6 @@ import Synchronization
   import SystemPackage
 #endif
 
-package struct SubprocessCommand: Sendable, Equatable {
-  package let arguments: [String]
-  package let timeout: Duration
-  package let captureLimit: Int
-  package let teardownGracePeriod: Duration
-  package let standardInput: Data
-  package let environmentKeysToRemove: [String]
-
-  package init(
-    arguments: [String],
-    timeout: Duration,
-    captureLimit: Int,
-    teardownGracePeriod: Duration,
-    standardInput: Data = Data(),
-    environmentKeysToRemove: [String] = []
-  ) {
-    precondition(captureLimit >= 0)
-    self.arguments = arguments
-    self.timeout = timeout
-    self.captureLimit = captureLimit
-    self.teardownGracePeriod = teardownGracePeriod
-    self.standardInput = standardInput
-    self.environmentKeysToRemove = environmentKeysToRemove
-  }
-}
-
-package enum SubprocessTermination: Sendable, Equatable {
-  case exited(Int32)
-  case signaled(Int32)
-  case timedOut
-  case cancelled
-  case startFailed(String)
-}
-
-package struct CapturedCommandStream: Sendable, Equatable {
-  package let bytes: Data
-  package let totalBytes: Int
-  package let truncated: Bool
-
-  package init(bytes: Data, totalBytes: Int, truncated: Bool) {
-    precondition(totalBytes >= bytes.count)
-    self.bytes = bytes
-    self.totalBytes = totalBytes
-    self.truncated = truncated
-  }
-}
-
-package struct SubprocessResult: Sendable, Equatable {
-  package let termination: SubprocessTermination
-
-  package let stdout: CapturedCommandStream
-  package let stderr: CapturedCommandStream
-
-  package let processIdentifier: Int32?
-
-  package init(
-    termination: SubprocessTermination,
-    stdout: CapturedCommandStream,
-    stderr: CapturedCommandStream,
-    processIdentifier: Int32?
-  ) {
-    self.termination = termination
-    self.stdout = stdout
-    self.stderr = stderr
-    self.processIdentifier = processIdentifier
-  }
-}
-
-package protocol SubprocessRunning: Sendable {
-  func run(_ command: SubprocessCommand) async -> SubprocessResult
-}
-
 package struct SwiftSubprocessRunner: SubprocessRunning {
   private let executablePath: String
   private let environmentForTesting: [String: String]
